"Change default rule value for item stack drop
The default rule value for the 'drop_item_stack_chance' was changed from 0 to -1 in the GraveGameRules and GraveUtils class. This change was made to correct the default function behavior of the rule where -1 more accurately signifies 'no chance' of item stack drop, in line with the intended behavior that should be initiated if this rule gets invoked.

Moreover, a conditional check was also updated within the GraveUtils class to match this change. Finally, an unnecessary duplication of item drop verification that was present has been removed from the codebase to improve efficiency and avoid repetition."